fix secret key
