+++ packages/app-mobile/src/screens/Unlocked/Settings/components/WalletList.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Alert, Pressable, StyleSheet, Text, View } from "react-native";
-import { Margin, WalletAddress } from "@components";
+import { Margin, WalletAddressLabel } from "@components";
 import type {
   // openPopupWindow,
   Blockchain,
@@ -129,12 +129,11 @@ export function WalletList({
 
 function AllWallets({ keys }) {
   return (
-    <>
+    <Margin horizontal={16} vertical={16}>
       {keys.map(({ name, publicKey, type }) => {
         return (
           <View key={publicKey.toString()}>
-            <Text>{JSON.stringify({ name, publicKey, type }, null, 2)}</Text>
-            <WalletAddress
+            <WalletAddressLabel
               name={name}
               publicKey={publicKey}
               style={{
@@ -149,7 +148,7 @@ function AllWallets({ keys }) {
           </View>
         );
       })}
-    </>
+    </Margin>
   );
 }
